pyproject.toml
[project]
name = "hackathon-backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Joaovictor23t4", email = "joaovictor239090@gmail.com"},
]
dependencies = ["setuptools>=80.9.0", "djangorestframework>=3.16.1", "django>=5.2.8", "drf-spectacular>=0.29.0", "djangorestframework-simplejwt>=5.5.1", "cloudinary>=1.44.1", "dotenv>=0.9.9", "pillow>=12.0.0", "dj-database-url>=3.0.1", "psycopg2-binary>=2.9.11", "django-filter>=25.2", "django-cors-headers>=4.9.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]

dev = "python manage.py runserver"
pre_migrate = "python manage.py makemigrations"
migrate = "python manage.py migrate"
startapp = "python manage.py startapp"
createsuperuser = "python manage.py createsuperuser"
shell = "python manage.py shell"
clean = "find . -name '*.pyc' -delete && find . -name '__pycache__' -exec rm -r {} +"
test = "python manage.py test"
lint = "ruff ."
format = "black ."
update = "pdm update"
install = "pdm install"
export-reqs = "pdm export -f requirements --without-hashes -o requirements.txt"